Added file diff for better change detection
