+++ go.mod
@@ -5,7 +5,6 @@ go 1.20
 require (
 	github.com/Masterminds/semver v1.5.0
 	github.com/fatih/color v1.16.0
-	github.com/go-resty/resty/v2 v2.10.0
 	github.com/gofrs/uuid v4.4.0+incompatible
 	github.com/google/go-github/v58 v58.0.0
 	github.com/hashicorp/hcl/v2 v2.19.1
@@ -20,7 +19,6 @@ require (
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mitchellh/go-wordwrap v1.0.1 // indirect
 	github.com/zclconf/go-cty v1.14.1 // indirect
-	golang.org/x/net v0.18.0 // indirect
 	golang.org/x/sys v0.14.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
 )
